renamed included .c files to .h
